+++ main.tf
@@ -12,7 +12,7 @@ resource "google_organization_iam_binding" "binding" {
   org_id = var.org_id
   role   = var.role
 
-  members = var.members
+  members = [for m in var.members : try(var.computed_members_map[regex("^computed:(.*)", m)[0]], m)]
 
   dynamic "condition" {
     for_each = var.condition != null ? ["condition"] : []
@@ -31,7 +31,7 @@ resource "google_organization_iam_member" "member" {
   org_id = var.org_id
   role   = var.role
 
-  member = each.value
+  member = try(var.computed_members_map[regex("^computed:(.*)", each.value)[0]], each.value)
 
   dynamic "condition" {
     for_each = var.condition != null ? ["condition"] : []
@@ -48,7 +48,7 @@ resource "google_organization_iam_policy" "policy" {
   count = var.module_enabled && local.create_policy ? 1 : 0
 
   org_id      = var.org_id
-  policy_data = data.google_iam_policy.policy[0].policy_data
+  policy_data = try(data.google_iam_policy.policy[0].policy_data, null)
 
   depends_on = [var.module_depends_on]
 }
@@ -61,7 +61,7 @@ data "google_iam_policy" "policy" {
 
     content {
       role    = binding.value.role
-      members = try(binding.value.members, var.members)
+      members = [for m in binding.value.members : try(var.computed_members_map[regex("^computed:(.*)", m)[0]], m)]
 
       dynamic "condition" {
         for_each = try([binding.value.condition], [])
@@ -74,6 +74,23 @@ data "google_iam_policy" "policy" {
       }
     }
   }
+
+  dynamic "audit_config" {
+    for_each = var.audit_configs
+
+    content {
+      service = audit_config.value.service
+
+      dynamic "audit_log_configs" {
+        for_each = audit_config.value.audit_log_configs
+
+        content {
+          log_type         = audit_log_configs.value.log_type
+          exempted_members = try(audit_log_configs.value.expected_members, null)
+        }
+      }
+    }
+  }
 }
 
 locals {
@@ -81,7 +98,7 @@ locals {
 }
 
 resource "google_organization_iam_audit_config" "organization" {
-  for_each = var.module_enabled ? local.audit_configs_map : {}
+  for_each = var.module_enabled && !local.create_policy ? local.audit_configs_map : {}
 
   org_id = var.org_id